requirements.txt
pandas==2.*
fire==0.6.*
colouredlogs==10.*
accelerate
tenacity==8.3.*
openai==1.30.*
torchvision
transformers==4.*
timm
ezcolorlog
diffusers[torch]
einops
flash_attn==2.6.3
decord
open_clip_torch
sentencepiece
python-dotenv
qwen_vl_utils